requirements.txt
# Core dependencies
pyyaml>=6.0
pydantic>=2.0
jsonschema>=4.20

# LLM integration
ollama>=0.1.0
langchain>=0.1.0
langchain-community>=0.0.20

# Data processing
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3

# RAG and embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0
pypdf>=3.17
langchain-text-splitters>=0.0.1

# Evaluation
rouge-score>=0.1.2
bert-score>=0.3.13

# Visualization
matplotlib>=3.7
seaborn>=0.12
plotly>=5.18

# Knowledge Graph (optional)
networkx>=3.1
rdflib>=7.0

# Utilities
tqdm>=4.66
python-dotenv>=1.0
tenacity>=8.2

# Development
pytest>=7.4
black>=23.0
